+++ noodles-gff/src/feature.rs
@@ -1,5 +1,6 @@
 //! Feature record.
 
 pub mod record;
+pub mod record_buf;
 
-pub use self::record::Record;
+pub use self::{record::Record, record_buf::RecordBuf};