+++ noodles-gff/src/feature/record_buf/attributes/field/value.rs
@@ -17,7 +17,7 @@ impl Value {
     /// # Examples
     ///
     /// ```
-    /// use noodles_gff::record_buf::attributes::field::Value;
+    /// use noodles_gff::feature::record_buf::attributes::field::Value;
     ///
     /// let value = Value::from("ndls");
     /// assert_eq!(value.as_string(), Some("ndls"));
@@ -37,7 +37,7 @@ impl Value {
     /// # Examples
     ///
     /// ```
-    /// use noodles_gff::record_buf::attributes::field::Value;
+    /// use noodles_gff::feature::record_buf::attributes::field::Value;
     ///
     /// let raw_values = vec![String::from("ndls0"), String::from("ndls1")];
     /// let value = Value::from(raw_values.clone());
@@ -58,7 +58,7 @@ impl Value {
     /// # Examples
     ///
     /// ```
-    /// use noodles_gff::record_buf::attributes::field::Value;
+    /// use noodles_gff::feature::record_buf::attributes::field::Value;
     ///
     /// let value = Value::from("ndls");
     /// let mut iter = value.iter();