+++ noodles-gff/src/io/reader.rs
@@ -13,7 +13,7 @@ use noodles_bgzf as bgzf;
 use noodles_core::Region;
 use noodles_csi::{self as csi, BinningIndex};
 
-use crate::{Line, RecordBuf};
+use crate::{feature::RecordBuf, Line};
 
 /// A GFF reader.
 pub struct Reader<R> {